+++ src/machdepArchConfigure.ml
@@ -9,13 +9,13 @@ let () =
     ]
   in
   C.main ~name:"model" ~args (fun c ->
-      let exe = "./machdep-ml" ^ !m ^ ".exe" in
+      let exe = "./machdep" ^ !m ^ "-ml.exe" in
       let {C.Process.exit_code; stdout; stderr} = C.Process.run c !real_gcc ["-D_GNUCC"; "-m" ^ !m; "machdep-ml.c"; "-o"; exe] in
       if exit_code = 0 then (
-        let {C.Process.stdout; stderr; exit_code} = C.Process.run c exe ["--env"] in
+        let {C.Process.stdout; stderr; exit_code} = C.Process.run c exe [] in
         assert (exit_code = 0);
-        print_string stdout;
+        Printf.printf "Some {%s}" stdout
       )
       else 
-        Printf.printf "nogcc%smodel\n" !m
+        Printf.printf "None"
     )