+++ test/small2/kernel1.c
@@ -1,4 +1,4 @@
-
+// This test is nonsense: DECLARE_WAIT_QUEUE_HEAD is a macro in Linux kernel
 DECLARE_WAIT_QUEUE_HEAD(log_wait);